refactor(RandomRaceController): RandomRaceController로 클래스 이름 수정
무작위 숫자 생성 로직과 전진 여부를 결정하는 Controller 로 클래스 이름을 변경한다.
